simplify: share the deviation arithmetic between the two verify scripts (tools-5)
ulp_distance and pair_dev were byte-identical in verify_sample.py and
verify_refs.py, and parse_wa was a third copy of the WANAMES/WAVALS scan.
verify_sample now imports pair_dev and parse_wa from verify_refs, the file
CI runs, so the recorder and the replay cannot drift. verify_refs keeps its
inline WA scan, which needs the names for its WA_FAIL detail.

Gate: verify_refs 119/119 within 1e-09, worst hmm_gaussian 3.63e-12 (35180
ulp), unchanged; verify_sample on eight_schools_noncentered VERIFIED at 0
ulp with docs/ regenerated identical.
